% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenoPlot.R
\name{allGenoPlot}
\alias{allGenoPlot}
\title{Helper function for plotting probabilities for all genotypes}
\usage{
allGenoPlot(markers, map, parents, title)
}
\description{
Helper function for plotting probabilities for all genotypes
}
\keyword{internal}
