% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBD.R
\name{c.calcIBD}
\alias{c.calcIBD}
\title{Concatenate function for objects of class calcIBD.}
\usage{
\method{c}{calcIBD}(...)
}
\arguments{
\item{...}{Objects of class \code{\link{calcIBD}}}
}
\value{
An object of class \code{\link{calcIBD}} containing data for all
concatenated objects.
}
\description{
Concatenates objects of class \code{\link{calcIBD}}. All objects that are
concatenated  should have the same population type and the same map. The
function is mainly meant for combining information for multiple crosses
with overlapping parents.
}
