% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcIBD}
\alias{calcIBD}
\title{Calculate IBD probabilities}
\usage{
calcIBD(poptype, locfile, mapfile, evalposfile = NULL, evaldist = NULL)
}
\arguments{
\item{poptype}{A character string indicating the type of population. One of
....}

\item{locfile}{A character string indicating the .loc file for the
population. The file should be in Flapjack format.}

\item{mapfile}{A character string indicating the .map file for the
population. The file should be in Flapjack format.}

\item{evalposfile}{An optional character string indicating a tab separated
.txt file containing combinations of chromosomes and positions to which the
calculations should be limited.}

\item{evaldist}{An optional numerical value indicating the maximum
distance for in between marker evaluation positions.}
}
\value{
A data.frame with IBD probabilities.
}
\description{
A long description
}
