% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbs.R
\name{getProbs}
\alias{getProbs}
\title{Extract Probabilities for markers}
\usage{
getProbs(IBD, markers)
}
\arguments{
\item{IBD}{An object of class calcIBD.}

\item{markers}{A character vector of markers that should be extracted.}
}
\value{
A data.frame with IBD probabilities for the extracted markers in the
column and genotypes in the rows.

@examples
\subsection{Compute IBD probabilities for Steptoe Morex.}{

SxMIBD <- calcIBD(poptype = "DH",
locfile = system.file("extdata/SxM", "SxM_geno.txt",
package = "statgenIBD"),
mapfile = system.file("extdata/SxM", "SxM_map.txt",
package = "statgenIBD"))
}

\subsection{Get probabilities for a single marker.}{

probOne <- getProbs(IBD = SxMIBD,
markers = "plc")
head(probOne)
}

\subsection{Get probabilities for a multiple markers.}{

probMult <- getProbs(IBD = SxMIBD,
markers = c("plc", "tuba1"))
head(probMult)
}
}
\description{
Extract IBD probabilities one or more markers from an object of class
calcIBD.
}
