% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQTLProb.R
\name{getQTLProb}
\alias{getQTLProb}
\title{Extract Probabilities for QTLS}
\usage{
getQTLProb(IBD, QTLS)
}
\arguments{
\item{IBD}{An object of class calcIBD.}

\item{QTLS}{A character vector of QTLS that should be extracted.}
}
\value{
A data.frame with IBD probabilities for the extracted QTLS in the
column and genotypes in the rows.

@examples
\subsection{Compute IBD probabilities for Steptoe Morex.}{

SxMIBD <- calcIBD(poptype = "DH",
locfile = system.file("extdata/SxM", "SxM_geno.txt",
package = "statgenIBD"),
mapfile = system.file("extdata/SxM", "SxM_map.txt",
package = "statgenIBD"))
}

\subsection{Get probabilities for a single QTL.}{

probOne <- getQTLProb(IBD = SxMIBD,
QTLS = "plc")
head(probOne)
}

\subsection{Get probabilities for a multiple QTLs.}{

probMult <- getQTLProb(IBD = SxMIBD,
QTLS = c("plc", "tuba1"))
head(probMult)
}
}
\description{
Extract IBD probabilities one or more QTLS from an object of class calcIBD.
}
