% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedPlot.R
\name{pedPlot}
\alias{pedPlot}
\title{Helper function for plotting the pedigree}
\usage{
pedPlot(pedigree, offSpring, popType, multiCross, genoCross, title)
}
\description{
Helper function for plotting the pedigree
}
\keyword{internal}
