% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBD.R
\name{plot.calcIBD}
\alias{plot.calcIBD}
\title{Plot function for objects of class calcIBD}
\usage{
\method{plot}{calcIBD}(x, ..., genotype, title = NULL, output = TRUE)
}
\arguments{
\item{x}{An object of class \code{GWAS}.}

\item{...}{Further arguments. Unused.}

\item{title}{A character string, the title of the plot.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE}, only a ggplot object is invisibly returned.}
}
\description{
Creates a plot for an object of class calcIBD.
}
