% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simQTL}
\alias{simQTL}
\title{simQTL}
\usage{
simQTL(
  inputfile,
  inbFnd,
  dir_name = NULL,
  mapfile = "map.txt",
  chrLength = 100,
  nlocChr = 11,
  nr_alleles = 2L,
  QTLPos = NULL,
  nrep = 1L,
  print_flexQTL = FALSE,
  fr_miss = 0,
  mu = 5,
  sigma2_e = 1,
  start_seed = 1234L
)
}
\description{
simQTL
}
