% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleGenoPlot.R
\name{singleGenoPlot}
\alias{singleGenoPlot}
\title{Helper function for plotting probabilities for a single genotype}
\usage{
singleGenoPlot(markers, map, parents, genotype, title)
}
\description{
Helper function for plotting probabilities for a single genotype
}
\keyword{internal}
