% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBD.R
\name{summary.calcIBD}
\alias{summary.calcIBD}
\title{Summary function for objects of class calcIBD}
\usage{
\method{summary}{calcIBD}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{calcIBD}}.}

\item{...}{Not used.}
}
\value{
No return value, a summary is printed.
}
\description{
Prints a short summary for objects of class \code{\link{calcIBD}}. The
summary consists of the population type, number of evaluation points,
number of individuals and names of the parents in the object.
}
\examples{
## Compute IBD probabilities for Steptoe Morex.
SxMIBD <- calcIBD(poptype = "DH",
                  locfile = system.file("extdata/SxM", "SxM_geno.txt",
                                        package = "statgenIBD"),
                  mapfile = system.file("extdata/SxM", "SxM_map.txt",
                                        package = "statgenIBD"))

## Print summary
summary(SxMIBD)

}
