% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDprob.R
\name{summary.IBDprob}
\alias{summary.IBDprob}
\title{Summary function for objects of class IBDprob}
\usage{
\method{summary}{IBDprob}(object, ...)
}
\arguments{
\item{object}{An object of class \code{IBDprob}.}

\item{...}{Not used.}
}
\value{
No return value, a summary is printed.
}
\description{
Prints a short summary for objects of class \code{IBDprob}. The
summary consists of the population type, number of evaluation points,
number of individuals and names of the parents in the object.
}
\examples{
## Compute IBD probabilities for Steptoe Morex.
SxMIBD <- calcIBD(popType = "DH",
                  markerFile = system.file("extdata/SxM", "SxM_geno.txt",
                                          package = "statgenIBD"),
                  mapFile = system.file("extdata/SxM", "SxM_map.txt",
                                        package = "statgenIBD"))

## Print summary
summary(SxMIBD)

}
