% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePedigree.R
\name{writePedigree}
\alias{writePedigree}
\title{Write pedigree to file.}
\usage{
writePedigree(IBDprob, outFile)
}
\arguments{
\item{IBDprob}{An object of class \code{IBDprob} containing the IBD
probabilities.}

\item{outFile}{A character string specifying the path of the output file.}
}
\value{
No output. The output file is created as a result of calling this
function.
}
\description{
Writes pedigree information to a plain text, tab-delimited file.
}
\examples{
## Compute IBD probabilities for Steptoe Morex.
SxMIBD <- calcIBD(popType = "DH",
                 markerFile = system.file("extdata/SxM", "SxM_geno.txt",
                                         package = "statgenIBD"),
                 mapFile = system.file("extdata/SxM", "SxM_map.txt",
                                      package = "statgenIBD"))

## Write predigree information to temporary file.
writePedigree(IBDprob = SxMIBD, outFile = tempfile(fileext = ".txt"))

}
