% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcIBD}
\alias{calcIBD}
\title{Calculate IBD probabilities}
\usage{
calcIBD(poptype, locfile, mapfile, evalposfile = NULL, evaldist = NULL)
}
\arguments{
\item{poptype}{A character string indicating the type of population. One of
DH, Fx, FxDH, BCx, BCxDH, BC1Sx, BC1SxDH, C3, C3DH, C3Sx, C3SxDH, C4, C4DH,
C4Sx, C4SxDH (see Details)}

\item{locfile}{A character string indicating the .loc file for the
population. The file should be in Flapjack format.}

\item{mapfile}{A character string indicating the .map file for the
population. The file should be in Flapjack format.}

\item{evalposfile}{An optional character string indicating a tab separated
.txt file containing combinations of chromosomes and positions to which the
calculations should be limited.}

\item{evaldist}{An optional numerical value indicating the maximum
distance for in between marker evaluation positions.}
}
\value{
An object of class \code{calcIBD}, a \code{list} with four elements,
\describe{
\item{map}{a \code{data.frame} with chromosome and position of the markers.}
\item{markers}{a 3-dimensional \code{array} of IBD probabilities with
markers, genotypes and  parents as array dimensions.}
\item{poptype}{the population type}
\item{multiCross}{a boolean indicating if multiple crosses have been
combined in the \code{calcIBD} object}
}
}
\description{
Calculate IBD probabilities for different types of populations.
}
\details{
IBD probabilities can be calculated for many different types of populations.
In the following table all supported populations are listed. Note that the
value of x in the population types is variable, with its maximum value
depicted in the last column.\tabular{llll}{
   \strong{Population type} \tab \strong{Cross} \tab \strong{Description} \tab \strong{max. x} \cr
   DH \tab biparental \tab doubled haploid population \tab  \cr
   Fx \tab biparental \tab Fx population (F1, followed by x-1 generations of selfing) \tab 8 \cr
   FxDH \tab biparental \tab Fx, followed by DH generation \tab 8 \cr
   BCx \tab biparental \tab backcross, second parent is recurrent parent \tab 9 \cr
   BCxDH \tab biparental \tab BCx, followed by DH generation \tab 9 \cr
   BC1Sx \tab biparental \tab BC1, followed by x generations of selfing \tab 7 \cr
   BC1SxDH \tab biparental \tab BC1, followed by x generations of selfing and DH \tab 6 \cr
   C3 \tab three-way \tab three way cross: (AxB) x C \tab  \cr
   C3DH \tab three-way \tab C3, followed by DH generation \tab  \cr
   C3Sx \tab three-way \tab C3, followed by x generation of selfing \tab 7 \cr
   C3SxDH \tab three-way \tab C3, followed by x generation of selfing and DH generation \tab 6 \cr
   C4 \tab four-way \tab four-way cross: (AxB) x (C x D) \tab  \cr
   C4DH \tab four-way \tab C4, followed by DH generation \tab  \cr
   C4Sx \tab four-way \tab C4, followed by x generation of selfing \tab 6 \cr
   C4SxDH \tab four-way \tab C4, followed by x generation of selfing and DH generation \tab 6 \cr
}
}
\examples{
## Compute IBD probabilities for Steptoe Morex.
SxMIBD <- calcIBD(poptype = "DH",
                  locfile = system.file("extdata", "SxM_geno.txt",
                                        package = "statgenIBD"),
                  mapfile = system.file("extdata", "SxM_map.txt",
                                        package = "statgenIBD"))

## Check summary
summary(SxMIBD)

}
