% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQTLProb.R
\name{getQTLProb}
\alias{getQTLProb}
\title{Extract Probabilities for QTLS}
\usage{
getQTLProb(IBD, QTLS)
}
\arguments{
\item{IBD}{An object of class calcIBD.}

\item{QTLS}{A character vector of QTLS that should be extracted.}
}
\value{
A data.frame with IBD probabilities for the extracted QTLS in the
column and genotypes in the rows.
}
\description{
Extract IBD probabilities one or more QTLS from an object of class calcIBD.
}
