% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDprob.R
\name{plot.IBDprob}
\alias{plot.IBDprob}
\title{Plot function for objects of class IBDprob}
\usage{
\method{plot}{IBDprob}(x, ..., genotype, title = genotype, output = TRUE)
}
\arguments{
\item{x}{An object of class \code{IBDprob}.}

\item{...}{Further arguments. Unused.}

\item{genotype}{A character string indicating the genotype for which the
plot should be made.}

\item{title}{A character string, the title of the plot.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE}, only a ggplot object is invisibly returned.}
}
\value{
A ggplot object is invisibly returned.
}
\description{
Creates a plot for an object of class \code{IBDprob}.
}
