% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDprob.R
\name{plot.IBDprob}
\alias{plot.IBDprob}
\title{Plot function for objects of class IBDprob}
\usage{
\method{plot}{IBDprob}(
  x,
  ...,
  plotType = c("singleGeno", "allGeno"),
  genotype,
  title = NULL,
  output = TRUE
)
}
\arguments{
\item{x}{An object of class \code{IBDprob}.}

\item{...}{Further arguments. Unused.}

\item{plotType}{A character string indicating the type of plot that should
be made.}

\item{genotype}{A character string indicating the genotype for which the
plot should be made. Ignored if \code{plotType = "allGeno"}.}

\item{title}{A character string, the title of the plot.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE}, only a ggplot object is invisibly returned.}
}
\value{
A ggplot object is invisibly returned.
}
\description{
Creates a plot for an object of class \code{IBDprob}. Two types of plot can
be made, a plot for a single genotype showing the IBD probabilities for all
parents across the genome (\code{plotType = "singleGeno"}), or a plot showing
for all genotypes the probabilities of the parent with the highest
probability per marker (\code{plotType = "allGeno"}).
}
\examples{
## Compute IBD probabilities for Steptoe Morex.
## Add extra evaluation positions in between exiting marker positions
## to assure evaluation positions are at most 2 cM apart.
SxMIBD_Ext <- calcIBD(popType = "DH",
                      markerFile = system.file("extdata/SxM", "SxM_geno.txt",
                                              package = "statgenIBD"),
                      mapFile = system.file("extdata/SxM", "SxM_map.txt",
                                           package = "statgenIBD"),
                      evalDist = 2)

## Plot results for genotype dh005.
plot(SxMIBD_Ext,
     plotType = "singleGeno",
     genotype = "dh005")

## Plot results for all genotypes.
plot(SxMIBD_Ext,
     plotType = "allGeno")

}
