% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIBDs.R
\name{readIBDs}
\alias{readIBDs}
\title{Read IBD probabilities from file}
\usage{
readIBDs(inFile)
}
\arguments{
\item{inFile}{A character string specifying the path of the file.}
}
\value{
An object of class \code{IBDprob} containing data a data.frame with
the IBD probabilities.
}
\description{
Reads IBD probabilities from a plain text, tab-delimited file. Information
about the file format can be found in the vignette (
\code{vignette("IBDFileFormat", package = "statgenIBD")}).
}
\examples{
## Read IBD probabilities for Steptoe Morex.
SxMIBD <- readIBDs(system.file("extdata/SxM", "SxM_IBDs.ibd",
                               package = "statgenIBD"))

## Print summary.
summary(SxMIBD)

}
