% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenIBD.R
\docType{package}
\name{statgenIBD-package}
\alias{statgenIBD}
\alias{statgenIBD-package}
\title{statgenIBD: Calculation of IBD probabilities}
\description{
Calculation of IBD probabilities for biparental, three and four-way 
    crosses.
}
\author{
\strong{Maintainer}: Martin Boer \email{martin.boer@wur.nl}

Other contributors:
\itemize{
  \item Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} [contributor]
}

}
\keyword{internal}
