% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenIBD.R
\docType{package}
\name{statgenIBD-package}
\alias{statgenIBD}
\alias{statgenIBD-package}
\title{statgenIBD: Calculation of IBD Probabilities}
\description{
For biparental, three and four-way crosses Identity by Descent 
    (IBD) probabilities can be calculated using Hidden Markov Models and 
    inheritance vectors following Lander and Green
    (<https://www.jstor.org/stable/29713>) and Huang
    (<doi:10.1073/pnas.1100465108>). One of a series of statistical genetic 
    packages for streamlining the analysis of typical plant breeding experiments
    developed by Biometris.
}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Authors:
\itemize{
  \item Martin Boer \email{martin.boer@wur.nl}
}

}
\keyword{internal}
