% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFlapjack.R
\name{writeFlapjack}
\alias{writeFlapjack}
\title{Write to Flapjack format}
\usage{
writeFlapjack(x, outFileMap = "ibd_map.txt", outFileGeno = "ibd_geno.txt")
}
\arguments{
\item{x}{An object of class \code{\link{IBD}}.}

\item{outFileMap}{A character string, the full path to the output map file.}

\item{outFileGeno}{A character string, the full path to the output genotype
file.}
}
\description{
Export the results of an IBD calculation to flapjack format so it can be
visualalized there.
}
